expose algorithms for selection in UI
